backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
playwright==1.48.0
google-genai>=1.0.0
python-dotenv==1.0.1
pydantic-settings==2.5.2
